pyproject.toml
[project]
name = "math-func-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.2.8",
    "python-decouple>=3.8",
]
